refactor: move main logic to Import class   - moves most of the code to feed_ursus/importer, which uses a new Importer class to store configuration (like the solr client, controlled fields, or collection names   - leaves click stuff in feed_ursus.py   - removes sinai mapping, that's got it's own feed_sinai repo
