+++ feed_ursus/fields/iiif_text_direction.yml
