+++ pyproject.toml
@@ -22,6 +22,9 @@ ipdb = "^0.13.13"
 black = {version = "^24.10.0", python = "^3.9"}
 flake8 = {version = "^7.1.1", python = "^3.9"}
 pylint = {version = "^3.3.1", python = "^3.9"}
+types-requests = "^2.32.0.20250602"
+types-pyyaml = "^6.0.12.20250516"
+requests-mock = "^1.12.1"
 
 [tool.poetry.group.test.dependencies]
 pytest = "*"
